moved finnhub lib file
